Move route handlers into their own files
