+++ demos/android-webview/app/src/main/assets/procaptcha.html
@@ -61,7 +61,7 @@
         </div>
         <div class="row">
             <div class="procaptcha w-300 m-auto"
-                 data-sitekey="5FNFBC97JQoJu7LBapycXbT66N9xzQH2tyFv1yNvBxqa8tQR"
+                 data-sitekey="5HDnHtTMEfqUGvRvC4SrDEyXk1HVX6vfTq6z9mpvfEggfnje"
                  data-captcha-type="image"></div>
         </div>
         <div class="row text-center">
