+++ packages/provider/src/api/captcha.ts
@@ -13,43 +13,14 @@
 // limitations under the License.
 
 import { handleErrors } from "@prosopo/api-express-router";
-import { ProsopoApiError } from "@prosopo/common";
-import { parseCaptchaAssets } from "@prosopo/datasets";
-import {
-	ApiParams,
-	type Captcha,
-	CaptchaRequestBody,
-	type CaptchaRequestBodyTypeOutput,
-	type CaptchaResponseBody,
-	CaptchaSolutionBody,
-	type CaptchaSolutionBodyType,
-	type CaptchaSolutionResponse,
-	CaptchaType,
-	ClientApiPaths,
-	type DappUserSolutionResult,
-	GetFrictionlessCaptchaChallengeRequestBody,
-	GetPowCaptchaChallengeRequestBody,
-	type GetPowCaptchaChallengeRequestBodyTypeOutput,
-	type GetPowCaptchaResponse,
-	type PowCaptchaSolutionResponse,
-	type ProsopoCaptchaCountConfigSchemaOutput,
-	SubmitPowCaptchaSolutionBody,
-	type SubmitPowCaptchaSolutionBodyTypeOutput,
-} from "@prosopo/types";
-import type { ScoreComponents } from "@prosopo/types-database";
+import { ClientApiPaths } from "@prosopo/types";
 import type { ProviderEnvironment } from "@prosopo/types-env";
-import { flatten, getIPAddress } from "@prosopo/util";
 import express, { type Router } from "express";
-import { getCompositeIpAddress } from "../compositeIpAddress.js";
-import { FrictionlessManager } from "../tasks/frictionless/frictionlessTasks.js";
-import { timestampDecayFunction } from "../tasks/frictionless/frictionlessTasksUtils.js";
-import { Tasks } from "../tasks/tasks.js";
-import { hashUserAgent } from "../utils/hashUserAgent.js";
-import { getMaintenanceMode } from "./admin/apiToggleMaintenanceModeEndpoint.js";
-import { getRequestUserScope } from "./blacklistRequestInspector.js";
-import { validateAddr, validateSiteKey } from "./validateAddress.js";
-
-const DEFAULT_FRICTIONLESS_THRESHOLD = 0.5;
+import getFrictionlessCaptchaChallenge from "./captcha/getFrictionlessCaptchaChallenge.js";
+import getImageCaptchaChallenge from "./captcha/getImageCaptchaChallenge.js";
+import getPoWCaptchaChallenge from "./captcha/getPoWCaptchaChallenge.js";
+import submitImageCaptchaSolution from "./captcha/submitImageCaptchaSolution.js";
+import submitPoWCaptchaSolution from "./captcha/submitPoWCaptchaSolution.js";
 
 /**
  * Returns a router connected to the database which can interact with the Proposo protocol
@@ -64,533 +35,30 @@ export function prosopoRouter(env: ProviderEnvironment): Router {
 
 	/**
 	 * Provides a Captcha puzzle to a Dapp User
-	 * @param {string} datasetId - Provider datasetId
-	 * @param {string} userAccount - Dapp User AccountId
-	 * @return {Captcha} - The Captcha data
 	 */
-	router.post(
-		ClientApiPaths.GetImageCaptchaChallenge,
-		async (req, res, next) => {
-			const tasks = new Tasks(env, req.logger);
-			let parsed: CaptchaRequestBodyTypeOutput;
-
-			if (!req.ip) {
-				return next(
-					new ProsopoApiError("API.BAD_REQUEST", {
-						context: { code: 400, error: "IP address not found" },
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-
-			const ipAddress = getIPAddress(req.ip || "");
-
-			try {
-				parsed = CaptchaRequestBody.parse(req.body);
-			} catch (err) {
-				return next(
-					new ProsopoApiError("CAPTCHA.PARSE_ERROR", {
-						context: { code: 400, error: err },
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-
-			const { datasetId, user, dapp, sessionId } = parsed;
-
-			validateSiteKey(dapp);
-			validateAddr(user);
-
-			try {
-				const clientRecord = await tasks.db.getClientRecord(dapp);
-
-				if (!clientRecord) {
-					return next(
-						new ProsopoApiError("API.SITE_KEY_NOT_REGISTERED", {
-							context: { code: 400, siteKey: dapp },
-							i18n: req.i18n,
-							logger: req.logger,
-						}),
-					);
-				}
-
-				const userScope = getRequestUserScope(
-					flatten(req.headers),
-					req.ja4,
-					req.ip,
-					user,
-				);
-				const userAccessPolicy = (
-					await tasks.imgCaptchaManager.getPrioritisedAccessPolicies(
-						userAccessRulesStorage,
-						dapp,
-						userScope,
-					)
-				)[0];
-
-				const {
-					valid,
-					reason,
-					sessionId: validSessionId,
-					solvedImagesCount,
-				} = await tasks.imgCaptchaManager.isValidRequest(
-					clientRecord,
-					CaptchaType.image,
-					env,
-					sessionId,
-					userAccessPolicy,
-					req.ip,
-				);
-
-				if (!valid) {
-					return next(
-						new ProsopoApiError(reason || "API.BAD_REQUEST", {
-							context: {
-								code: 400,
-								siteKey: dapp,
-								user,
-							},
-							i18n: req.i18n,
-							logger: req.logger,
-						}),
-					);
-				}
-
-				const captchaConfig: ProsopoCaptchaCountConfigSchemaOutput = {
-					solved: {
-						count:
-							solvedImagesCount ||
-							userAccessPolicy?.solvedImagesCount ||
-							env.config.captchas.solved.count,
-					},
-					unsolved: {
-						count:
-							userAccessPolicy?.unsolvedImagesCount ||
-							env.config.captchas.unsolved.count,
-					},
-				};
-
-				const taskData =
-					await tasks.imgCaptchaManager.getRandomCaptchasAndRequestHash(
-						datasetId,
-						user,
-						ipAddress,
-						captchaConfig,
-						clientRecord.settings.imageThreshold ?? 0.8,
-						validSessionId,
-					);
-				const captchaResponse: CaptchaResponseBody = {
-					[ApiParams.status]: "ok",
-					[ApiParams.captchas]: taskData.captchas.map((captcha: Captcha) => ({
-						...captcha,
-						target: req.t(`TARGET.${captcha.target}`),
-						items: captcha.items.map((item) =>
-							parseCaptchaAssets(item, env.assetsResolver),
-						),
-					})),
-					[ApiParams.requestHash]: taskData.requestHash,
-					[ApiParams.timestamp]: taskData.timestamp.toString(),
-					[ApiParams.signature]: {
-						[ApiParams.provider]: {
-							[ApiParams.requestHash]: taskData.signedRequestHash,
-						},
-					},
-				};
-				req.logger.info(() => ({
-					msg: "Image captcha challenge issued",
-					data: {
-						captchaType: CaptchaType.image,
-						requestHash: taskData.requestHash,
-						solvedImagesCount: captchaConfig.solved.count,
-						user,
-						dapp,
-						sessionId,
-					},
-				}));
-				return res.json(captchaResponse);
-			} catch (err) {
-				req.logger.error(() => ({
-					err,
-					data: req.params,
-					msg: "Error in image captcha challenge request",
-				}));
-				return next(
-					new ProsopoApiError("API.BAD_REQUEST", {
-						context: {
-							error: err,
-
-							code: 500,
-							params: req.params,
-							context: err,
-						},
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-		},
+	router.post(ClientApiPaths.GetImageCaptchaChallenge, (req, res, next) =>
+		getImageCaptchaChallenge(env, userAccessRulesStorage)(req, res, next),
 	);
 
 	/**
 	 * Receives solved CAPTCHA challenges from the user, stores to database, and checks against solution commitment
-	 *
-	 * @param {string} userAccount - Dapp User id
-	 * @param {string} dappAccount - Dapp Contract AccountId
-	 * @param {Captcha[]} captchas - The Captcha solutions
-	 * @return {DappUserSolutionResult} - The Captcha solution result and proof
 	 */
-	router.post(
-		ClientApiPaths.SubmitImageCaptchaSolution,
-		async (req, res, next) => {
-			const tasks = new Tasks(env, req.logger);
-
-			// If in maintenance mode, always return verified
-			if (getMaintenanceMode()) {
-				req.logger.info(() => ({
-					msg: "Maintenance mode active - returning verified for image captcha",
-				}));
-				const result: CaptchaSolutionResponse = {
-					status: "ok",
-					captchas: [],
-					verified: true,
-				};
-				return res.json(result);
-			}
-
-			let parsed: CaptchaSolutionBodyType;
-			try {
-				parsed = CaptchaSolutionBody.parse(req.body);
-			} catch (err) {
-				return next(
-					new ProsopoApiError("CAPTCHA.PARSE_ERROR", {
-						context: { code: 400, error: err, body: req.body },
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-
-			const { user, dapp } = parsed;
-
-			validateSiteKey(dapp);
-			validateAddr(user);
-
-			try {
-				const clientRecord = await tasks.db.getClientRecord(parsed.dapp);
-
-				if (!clientRecord) {
-					return next(
-						new ProsopoApiError("API.SITE_KEY_NOT_REGISTERED", {
-							context: { code: 400, siteKey: dapp },
-							i18n: req.i18n,
-							logger: req.logger,
-						}),
-					);
-				}
-
-				const result: DappUserSolutionResult =
-					await tasks.imgCaptchaManager.dappUserSolution(
-						user,
-						dapp,
-						parsed[ApiParams.requestHash],
-						parsed[ApiParams.captchas],
-						parsed[ApiParams.signature].user.timestamp,
-						Number.parseInt(parsed[ApiParams.timestamp]),
-						parsed[ApiParams.signature].provider.requestHash,
-						getIPAddress(req.ip || ""),
-						flatten(req.headers),
-						req.ja4,
-					);
-
-				const returnValue: CaptchaSolutionResponse = {
-					status: req.i18n.t(
-						result.verified ? "API.CAPTCHA_PASSED" : "API.CAPTCHA_FAILED",
-					),
-					...result,
-				};
-				return res.json(returnValue);
-			} catch (err) {
-				req.logger.error(() => ({
-					err,
-					body: req.body,
-					msg: "Error in image captcha solution submission",
-				}));
-				return next(
-					new ProsopoApiError("API.BAD_REQUEST", {
-						context: {
-							code: 500,
-							siteKey: req.body.dapp,
-							error: err,
-						},
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-		},
+	router.post(ClientApiPaths.SubmitImageCaptchaSolution, (req, res, next) =>
+		submitImageCaptchaSolution(env, userAccessRulesStorage)(req, res, next),
 	);
 
 	/**
 	 * Supplies a PoW challenge to a Dapp User
-	 *
-	 * @param {string} userAccount - User address
-	 * @param {string} dappAccount - Dapp address
 	 */
-	router.post(ClientApiPaths.GetPowCaptchaChallenge, async (req, res, next) => {
-		let parsed: GetPowCaptchaChallengeRequestBodyTypeOutput;
-		const tasks = new Tasks(env);
-		tasks.setLogger(req.logger);
-
-		try {
-			parsed = GetPowCaptchaChallengeRequestBody.parse(req.body);
-		} catch (err) {
-			return next(
-				new ProsopoApiError("CAPTCHA.PARSE_ERROR", {
-					context: { code: 400, error: err },
-					i18n: req.i18n,
-					logger: req.logger,
-				}),
-			);
-		}
-
-		const { user, dapp, sessionId } = parsed;
-
-		validateSiteKey(dapp);
-		validateAddr(user);
-
-		try {
-			const clientSettings = await tasks.db.getClientRecord(dapp);
-
-			if (!clientSettings) {
-				return next(
-					new ProsopoApiError("API.SITE_KEY_NOT_REGISTERED", {
-						context: { code: 400, siteKey: dapp },
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-
-			const userScope = getRequestUserScope(
-				flatten(req.headers),
-				req.ja4,
-				req.ip,
-				user,
-			);
-			const userAccessPolicy = (
-				await tasks.powCaptchaManager.getPrioritisedAccessPolicies(
-					userAccessRulesStorage,
-					dapp,
-					userScope,
-				)
-			)[0];
-
-			const {
-				valid,
-				reason,
-				sessionId: validSessionId,
-				powDifficulty,
-			} = await tasks.powCaptchaManager.isValidRequest(
-				clientSettings,
-				CaptchaType.pow,
-				env,
-				sessionId,
-				userAccessPolicy,
-				req.ip,
-			);
-
-			if (!valid) {
-				return next(
-					new ProsopoApiError(reason || "API.BAD_REQUEST", {
-						context: {
-							code: 400,
-							siteKey: dapp,
-							user,
-						},
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-
-			const origin = req.headers.origin;
-
-			if (!origin) {
-				return next(
-					new ProsopoApiError("API.BAD_REQUEST", {
-						context: {
-							error: "Origin header not found",
-							code: 400,
-							siteKey: dapp,
-							user,
-						},
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-
-			const difficulty =
-				powDifficulty ||
-				userAccessPolicy?.powDifficulty ||
-				clientSettings?.settings?.powDifficulty;
-			const challenge = await tasks.powCaptchaManager.getPowCaptchaChallenge(
-				user,
-				dapp,
-				origin,
-				difficulty,
-			);
-
-			await tasks.db.storePowCaptchaRecord(
-				challenge.challenge,
-				{
-					requestedAtTimestamp: challenge.requestedAtTimestamp,
-					userAccount: user,
-					dappAccount: dapp,
-				},
-				challenge.difficulty,
-				challenge.providerSignature,
-				getCompositeIpAddress(req.ip || ""),
-				flatten(req.headers),
-				req.ja4,
-				validSessionId,
-			);
-
-			const getPowCaptchaResponse: GetPowCaptchaResponse = {
-				[ApiParams.status]: "ok",
-				[ApiParams.challenge]: challenge.challenge,
-				[ApiParams.difficulty]: challenge.difficulty,
-				[ApiParams.timestamp]: challenge.requestedAtTimestamp.toString(),
-				[ApiParams.signature]: {
-					[ApiParams.provider]: {
-						[ApiParams.challenge]: challenge.providerSignature,
-					},
-				},
-			};
-
-			req.logger.info(() => ({
-				msg: "PoW captcha challenge issued",
-				data: {
-					captchaType: CaptchaType.pow,
-					challenge: challenge.challenge,
-					difficulty: challenge.difficulty,
-					user,
-					dapp,
-					session: sessionId,
-				},
-			}));
-			return res.json(getPowCaptchaResponse);
-		} catch (err) {
-			req.logger.error(() => ({
-				err,
-				body: req.body,
-				msg: "Error in PoW captcha challenge request",
-			}));
-			return next(
-				new ProsopoApiError("API.BAD_REQUEST", {
-					context: {
-						code: 500,
-						siteKey: req.body.dapp,
-						user: req.body.user,
-						error: err,
-					},
-					i18n: req.i18n,
-					logger: req.logger,
-				}),
-			);
-		}
-	});
+	router.post(ClientApiPaths.GetPowCaptchaChallenge, (req, res, next) =>
+		getPoWCaptchaChallenge(env, userAccessRulesStorage)(req, res, next),
+	);
 
 	/**
 	 * Verifies a user's PoW solution as being approved or not
-	 *
-	 * @param {string} challenge - the challenge string
-	 * @param {number} difficulty - the difficulty of the challenge
-	 * @param {string} signature - the signature of the challenge
-	 * @param {string} nonce - the nonce of the challenge
-	 * @param {number} verifiedTimeout - the valid length of captcha solution in ms
 	 */
-	router.post(
-		ClientApiPaths.SubmitPowCaptchaSolution,
-		async (req, res, next) => {
-			let parsed: SubmitPowCaptchaSolutionBodyTypeOutput;
-			const tasks = new Tasks(env, req.logger);
-
-			// If in maintenance mode, always return verified
-			if (getMaintenanceMode()) {
-				req.logger.info(() => ({
-					msg: "Maintenance mode active - returning verified",
-				}));
-				const response: PowCaptchaSolutionResponse = {
-					status: "ok",
-					verified: true,
-				};
-				return res.json(response);
-			}
-
-			try {
-				parsed = SubmitPowCaptchaSolutionBody.parse(req.body);
-			} catch (err) {
-				return next(
-					new ProsopoApiError("CAPTCHA.PARSE_ERROR", {
-						context: { code: 400, error: err, body: req.body },
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-
-			const { challenge, signature, nonce, verifiedTimeout, dapp, user } =
-				parsed;
-
-			validateSiteKey(dapp);
-			validateAddr(user);
-
-			try {
-				const clientRecord = await tasks.db.getClientRecord(dapp);
-
-				if (!clientRecord) {
-					return next(
-						new ProsopoApiError("API.SITE_KEY_NOT_REGISTERED", {
-							context: { code: 400, siteKey: dapp },
-							i18n: req.i18n,
-							logger: req.logger,
-						}),
-					);
-				}
-
-				const verified = await tasks.powCaptchaManager.verifyPowCaptchaSolution(
-					challenge,
-					signature.provider.challenge,
-					nonce,
-					verifiedTimeout,
-					signature.user.timestamp,
-					getIPAddress(req.ip || ""),
-					flatten(req.headers),
-				);
-				const response: PowCaptchaSolutionResponse = { status: "ok", verified };
-				return res.json(response);
-			} catch (err) {
-				req.logger.error(() => ({
-					err,
-					body: req.body,
-					msg: "Error in PoW captcha solution submission",
-				}));
-				return next(
-					new ProsopoApiError("API.BAD_REQUEST", {
-						context: {
-							code: 500,
-							siteKey: req.body.dapp,
-							error: err,
-						},
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-		},
+	router.post(ClientApiPaths.SubmitPowCaptchaSolution, (req, res, next) =>
+		submitPoWCaptchaSolution(env)(req, res, next),
 	);
 
 	/**
@@ -598,311 +66,12 @@ export function prosopoRouter(env: ProviderEnvironment): Router {
 	 */
 	router.post(
 		ClientApiPaths.GetFrictionlessCaptchaChallenge,
-		async (req, res, next) => {
-			try {
-				const tasks = new Tasks(env, req.logger);
-				const { token, headHash, dapp, user } =
-					GetFrictionlessCaptchaChallengeRequestBody.parse(req.body);
-
-				// If in maintenance mode, store dummy token and send PoW captcha
-				if (getMaintenanceMode()) {
-					req.logger.info(() => ({
-						msg: "Maintenance mode active - storing dummy token and sending PoW captcha",
-						data: { dapp, user },
-					}));
-
-					// Send PoW captcha with dummy frictionless data
-					return res.json(
-						await tasks.frictionlessManager.sendPowCaptcha({
-							token,
-							score: 0,
-							threshold: 0.5,
-							scoreComponents: {
-								baseScore: 0,
-							},
-							providerSelectEntropy: 0,
-							ipAddress: getCompositeIpAddress(req.ip || ""),
-							powDifficulty: undefined,
-							webView: false,
-							iFrame: false,
-							decryptedHeadHash: "",
-						}),
-					);
-				}
-
-				// Check if the token has already been used
-				const existingToken = await tasks.db.getSessionRecordByToken(token);
-
-				if (existingToken) {
-					req.logger.info(() => ({
-						token: existingToken,
-						msg: "Token has already been used",
-					}));
-					return next(
-						new ProsopoApiError("API.BAD_REQUEST", {
-							context: {
-								code: 400,
-								siteKey: dapp,
-								user,
-							},
-							i18n: req.i18n,
-							logger: req.logger,
-						}),
-					);
-				}
-
-				const lScore = tasks.frictionlessManager.checkLangRules(
-					req.headers["accept-language"] || "",
-				);
-
-				const {
-					baseBotScore,
-					timestamp,
-					providerSelectEntropy,
-					userId,
-					userAgent,
-					webView,
-					iFrame,
-					decryptedHeadHash,
-				} = await tasks.frictionlessManager.decryptPayload(token, headHash);
-
-				req.logger.debug(() => ({
-					msg: "Decrypted payload",
-					data: {
-						baseBotScore,
-						timestamp,
-						providerSelectEntropy,
-						userId,
-						userAgent,
-						webView,
-					},
-				}));
-
-				let botScore = baseBotScore + lScore;
-
-				const clientRecord = await tasks.db.getClientRecord(dapp);
-
-				if (!clientRecord) {
-					return next(
-						new ProsopoApiError("API.SITE_KEY_NOT_REGISTERED", {
-							context: { code: 400, siteKey: dapp },
-							i18n: req.i18n,
-							logger: req.logger,
-						}),
-					);
-				}
-
-				const { valid, reason } =
-					await tasks.frictionlessManager.isValidRequest(
-						clientRecord,
-						CaptchaType.frictionless,
-						env,
-					);
-
-				if (!valid) {
-					return next(
-						new ProsopoApiError(reason || "API.BAD_REQUEST", {
-							context: {
-								code: 400,
-								siteKey: dapp,
-								user,
-							},
-							i18n: req.i18n,
-							logger: req.logger,
-						}),
-					);
-				}
-
-				const botThreshold =
-					clientRecord.settings?.frictionlessThreshold ||
-					DEFAULT_FRICTIONLESS_THRESHOLD;
-
-				// Initialize score components
-				let scoreComponents: ScoreComponents = {
-					baseScore: baseBotScore,
-					...(lScore && { lScore }),
-				};
-
-				const ipAddress = getCompositeIpAddress(req.ip || "");
-
-				// Set common session parameters on the frictionless manager
-				tasks.frictionlessManager.setSessionParams({
-					token,
-					score: botScore,
-					threshold: botThreshold,
-					scoreComponents,
-					providerSelectEntropy,
-					ipAddress,
-					webView,
-					iFrame,
-					decryptedHeadHash,
-				});
-
-				// Check if the IP address is blocked
-				const userScope = getRequestUserScope(
-					flatten(req.headers),
-					req.ja4,
-					req.ip,
-					user,
-				);
-				const userAccessPolicy = (
-					await tasks.frictionlessManager.getPrioritisedAccessPolicies(
-						userAccessRulesStorage,
-						dapp,
-						userScope,
-					)
-				)[0];
-
-				// Check the user agent in token and user id in request match request
-				// Hash the request user agent to compare with the hashed user agent from the token
-				const headersUserAgent = req.headers["user-agent"];
-				const hashedHeadersUserAgent = headersUserAgent
-					? hashUserAgent(headersUserAgent)
-					: "";
-				const headersProsopoUser = req.headers["prosopo-user"];
-				if (
-					hashedHeadersUserAgent !== userAgent ||
-					headersProsopoUser !== userId
-				) {
-					req.logger.info(() => ({
-						msg: "User agent or user id does not match",
-						data: {
-							headersUserAgent,
-							hashedHeadersUserAgent,
-							userAgent: userAgent, // This is the hashed user agent from the token
-							headersProsopoUser,
-							userId,
-						},
-					}));
-					return res.json(
-						await tasks.frictionlessManager.sendImageCaptcha({
-							solvedImagesCount: timestampDecayFunction(timestamp),
-						}),
-					);
-				}
-
-				// If the user or IP address has an image captcha config defined, send an image captcha
-				if (userAccessPolicy) {
-					const scoreUpdate =
-						tasks.frictionlessManager.scoreIncreaseAccessPolicy(
-							userAccessPolicy,
-							baseBotScore,
-							botScore,
-							scoreComponents,
-						);
-					botScore = scoreUpdate.score;
-					scoreComponents = scoreUpdate.scoreComponents;
-					tasks.frictionlessManager.updateScore(botScore, scoreComponents);
-
-					if (userAccessPolicy.captchaType === CaptchaType.image) {
-						return res.json(
-							await tasks.frictionlessManager.sendImageCaptcha({
-								solvedImagesCount: userAccessPolicy.solvedImagesCount,
-							}),
-						);
-					}
-					if (userAccessPolicy.captchaType === CaptchaType.pow) {
-						return res.json(
-							await tasks.frictionlessManager.sendPowCaptcha({
-								powDifficulty: undefined,
-							}),
-						);
-					}
-				}
-
-				if (clientRecord.settings.disallowWebView && webView) {
-					tasks.logger.info(() => ({
-						msg: "WebView detected",
-					}));
-					const scoreUpdate = tasks.frictionlessManager.scoreIncreaseWebView(
-						baseBotScore,
-						botScore,
-						scoreComponents,
-					);
-					botScore = scoreUpdate.score;
-					scoreComponents = scoreUpdate.scoreComponents;
-					tasks.frictionlessManager.updateScore(botScore, scoreComponents);
-
-					return res.json(
-						await tasks.frictionlessManager.sendImageCaptcha({
-							solvedImagesCount: env.config.captchas.solved.count * 2,
-						}),
-					);
-				}
-
-				// If the timestamp is older than 10 minutes, send an image captcha
-				if (FrictionlessManager.timestampTooOld(timestamp)) {
-					const scoreUpdate = tasks.frictionlessManager.scoreIncreaseTimestamp(
-						timestamp,
-						baseBotScore,
-						botScore,
-						scoreComponents,
-					);
-					botScore = scoreUpdate.score;
-					scoreComponents = scoreUpdate.scoreComponents;
-					tasks.frictionlessManager.updateScore(botScore, scoreComponents);
-
-					return res.json(
-						await tasks.frictionlessManager.sendImageCaptcha({
-							solvedImagesCount: timestampDecayFunction(timestamp),
-						}),
-					);
-				}
-
-				// If the host is not verified, send an image captcha
-				const hostVerified = await tasks.frictionlessManager.hostVerified(
-					providerSelectEntropy,
-				);
-				if (!hostVerified.verified) {
-					const scoreUpdate =
-						tasks.frictionlessManager.scoreIncreaseUnverifiedHost(
-							hostVerified.domain,
-							baseBotScore,
-							botScore,
-							scoreComponents,
-						);
-					botScore = scoreUpdate.score;
-					scoreComponents = scoreUpdate.scoreComponents;
-					tasks.frictionlessManager.updateScore(botScore, scoreComponents);
-				}
-
-				// If the bot score is greater than the threshold, send an image captcha
-				if (Number(botScore) > botThreshold) {
-					req.logger.info(() => ({
-						msg: "Bot score is greater than threshold",
-						data: {
-							botScore,
-							botThreshold,
-							token,
-						},
-					}));
-					return res.json(
-						await tasks.frictionlessManager.sendImageCaptcha({
-							solvedImagesCount: env.config.captchas.solved.count,
-						}),
-					);
-				}
-
-				// Otherwise, send a PoW captcha
-				return res.json(
-					await tasks.frictionlessManager.sendPowCaptcha({
-						powDifficulty: undefined,
-					}),
-				);
-			} catch (err) {
-				req.logger.error(() => ({
-					err,
-					msg: "Error in frictionless captcha challenge",
-				}));
-				return next(
-					new ProsopoApiError("API.BAD_REQUEST", {
-						context: { code: 400, error: err },
-						i18n: req.i18n,
-						logger: req.logger,
-					}),
-				);
-			}
-		},
+		(req, res, next) =>
+			getFrictionlessCaptchaChallenge(env, userAccessRulesStorage)(
+				req,
+				res,
+				next,
+			),
 	);
 
 	// Your error handler should always be at the end of your application stack. Apparently it means not only after all
